Add a test that crashes the Codegen.
This adds a test that crashes the Codegen, the schema-changes +
operations are based on code we've got in our repo now, that chokes the
new codegen.

The modeling of `BookPrice` might not be the best way to do it in
GraphQL, but it's similar to what we've got, and it has worked with the
old codegen.